requirements.txt
yt-dlp
openai-whisper
opencv-python
pytesseract
pillow
requests
google-generativeai